+++ src/main/ReactJS/README.md
@@ -1,0 +1,11 @@
+# **Team X-Men Spring 2022**
+### Open Source Tracking project for Commerce Bank
+-----------------------------------------------------
+**March 24, 2022**
+- Initial commit with template project
+
+**March 25, 2022** 
+- Add basic react parts
+
+**March 30, 2022**
+- Add Header and Footer
