+++ src/main/java/com/SpringReactApplication.java
@@ -1,4 +1,4 @@
-package com.example.SpringReact;
+package com;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
