+++ src/main/java/com/controller/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package com.example.SpringReact.controller;
+package com.controller;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
