+++ src/main/java/com/domain/Account.java
@@ -1,17 +1,21 @@
-package com.example.SpringReact.domain;
+package com.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.Entity;
+import javax.persistence.Id;
 
+@Data
 @AllArgsConstructor
 @NoArgsConstructor
-@Data
-public class Login {
+@Entity
+public class Account {
 
+    @Id
     private String name;
+
     private String password;
 
 }
