+++ src/main/java/com/domain/Login.java
@@ -1,22 +1,17 @@
-package com.example.SpringReact.domain;
+package com.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.Entity;
-import javax.persistence.Id;
-import java.util.Date;
 
-@Data
 @AllArgsConstructor
 @NoArgsConstructor
-@Entity
-public class Account {
+@Data
+public class Login {
 
-    @Id
     private String name;
-
     private String password;
 
 }
