+++ src/main/java/com/domain/OpenSourceProjects.java
@@ -1,4 +1,4 @@
-package com.example.SpringReact.domain;
+package com.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
