+++ src/main/java/com/domain/Token.java
@@ -1,4 +1,4 @@
-package com.example.SpringReact.domain;
+package com.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
