--- src/main/java/com/example/SpringReact/domain/BookRepository.java
@@ -1,11 +1,0 @@
-package com.example.SpringReact.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-
-@Repository
-public interface BookRepository extends JpaRepository<Book, Long>{
-
-
-}
