+++ src/main/java/com/repository/AccountRepository.java
@@ -1,7 +1,7 @@
-package com.example.SpringReact.repository;
+package com.repository;
 
-import com.example.SpringReact.domain.Account;
-import com.example.SpringReact.domain.User;
+import com.domain.Account;
+import com.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AccountRepository extends JpaRepository<Account,String> {
